pyproject.toml
[tool.poetry]
name = "zupit"
version = "0.1.0"
description = ""
authors = ["azevedo <azevedoantoniocicero@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg = {extras = ["binary"], version = "^3.2.1"}
jinja2 = "^3.1.4"
fastapi = {extras = ["standard"], version = "^0.112.0"}
pydantic-settings = "^2.4.0"
httpx = "^0.27.0"
email-validator = "^2.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.5.7"
httpx = "^0.27.0"
testcontainers = "^4.8.0"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'fastapi dev zupit/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=zupit -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
